Add partially done broadcasting scenario
The broadcasting still doesn't work, because of an unexpected issue:
Since the PerfectLinkHandler only handles client connections which come and go, its state is erased often. Therefore, the data that is stored in its state such as the process id structures and its own process id struct are erased. One idea would be to send this data to a PerfectLinkLayerMemory process that is specifically designed to store state. The reason why we have to do this is to keep the server asynchronous, we cannot really avoid using the PerfectLinkHandler as it is in this situation.
To fix the issue, the PerfectLinkLayer needs to spawn a separate custom GenServer process, let's name it "PerfectLinkLayerMemory". This custom GenServer will store our state information and will send and receive data to/from the PerfectLinkHandler. When initiating a PerfectLinkHandler instance, the PID of the PerfectLinkLayerMemory will be passed to it, thus enabling interprocess communication between the PerfectLinkHandler and the PerfectLinkLayerMemory.