Create a process that holds state for a node
The PerfectLinkLayerMemory process is supposed to store the state elements such as the process ID structures received from the hub when the system is initialized. This approach avoids the issue of the state of the process spawned through the DynamicSupervisor being thrown away/deleted, instead the data received at the system initialization stage is sent by the PerfectLinkHandler to the PerfectLinkLayerMemory process and these two communicate whenever it is necessary.

TODO: The broadcast mechanism is still not done, as there are errors
related to the message types. Apparently, the reference processes of the
hub app send an APP_VALUE message to the other processes when required
to broadcast the message forwards.