Handle acknowledgment values
One issue is that the AppWrite message received by the leader process might have to be broadcasted to all the other processes. Afterwards, the same process should unfold. The reason for this assumption is that the processes other than the leader process cannot know what the register name is and the AppWrite message is the only message that contains it and which could be broadcasted to the other processes.