Fix instant heartbeat error
The app is still not working, but heartbeat requests and replies seem to be properly interpreted by the hub and other instances. This is assumed because heartbeat replies are actually received by the request sender, however those might just be the processes of this project and not the ones of the reference binary. Abstraction IDs still do not work.