Fix abstraction IDs in the UC/EP layers
The project still fails in the EPFD layer because the initialization function tries to access the delay variable which is actually located in the consensus_dictionary map containing the consensus_topic_data. Abstraction IDs and state management still have to be verified and updated accordingly in the EPFD/ELD layers.