Fix EMFILE errors
The errors were related to the maximum limit of file descriptors on the operating system. To fix this, the ulimit utility can be used with the -n option to set the limit to a higher value. In this case, that value was 8192 (chosen randomly).

There are still bugs possibly related to mismatched Protobuf types.